src/requirements.txt
fastapi==0.115.5
langchain==0.3.7
langchain_community==0.3.7
langchain_core==0.3.19
langchain_openai==0.2.9
langgraph==0.2.53
pandas==2.2.3
pydantic==2.10.0
python-dotenv==1.0.1
streamlit==1.40.1
uvicorn==0.32.1
langgraph-checkpoint==2.0.10
duckduckgo_search==8.0.2